+++ obs-studio-client/source/simple-replay-buffer.cpp
@@ -66,7 +66,7 @@ osn::SimpleReplayBuffer::SimpleReplayBuffer(const Napi::CallbackInfo &info) : Na
 {
 	Napi::Env env = info.Env();
 	Napi::HandleScope scope(env);
-	int length = info.Length();
+	size_t length = info.Length();
 
 	if (length <= 0 || !info[0].IsNumber()) {
 		Napi::TypeError::New(env, "Number expected").ThrowAsJavaScriptException();
@@ -88,7 +88,7 @@ Napi::Value osn::SimpleReplayBuffer::Create(const Napi::CallbackInfo &info)
 	if (!ValidateResponse(info, response))
 		return info.Env().Undefined();
 
-	auto instance = osn::SimpleReplayBuffer::constructor.New({Napi::Number::New(info.Env(), response[1].value_union.ui64)});
+	auto instance = osn::SimpleReplayBuffer::constructor.New({Napi::Number::New(info.Env(), static_cast<double>(response[1].value_union.ui64))});
 
 	return instance;
 }
@@ -124,7 +124,7 @@ Napi::Value osn::SimpleReplayBuffer::GetLegacySettings(const Napi::CallbackInfo 
 	if (!ValidateResponse(info, response))
 		return info.Env().Undefined();
 
-	auto instance = osn::SimpleReplayBuffer::constructor.New({Napi::Number::New(info.Env(), response[1].value_union.ui64)});
+	auto instance = osn::SimpleReplayBuffer::constructor.New({Napi::Number::New(info.Env(), static_cast<double>(response[1].value_union.ui64))});
 
 	return instance;
 }
@@ -159,7 +159,7 @@ Napi::Value osn::SimpleReplayBuffer::GetStreaming(const Napi::CallbackInfo &info
 	if (!ValidateResponse(info, response))
 		return info.Env().Undefined();
 
-	auto instance = osn::SimpleStreaming::constructor.New({Napi::Number::New(info.Env(), response[1].value_union.ui64)});
+	auto instance = osn::SimpleStreaming::constructor.New({Napi::Number::New(info.Env(), static_cast<double>(response[1].value_union.ui64))});
 	return instance;
 }
 
@@ -190,7 +190,7 @@ Napi::Value osn::SimpleReplayBuffer::GetRecording(const Napi::CallbackInfo &info
 	if (!ValidateResponse(info, response))
 		return info.Env().Undefined();
 
-	auto instance = osn::SimpleRecording::constructor.New({Napi::Number::New(info.Env(), response[1].value_union.ui64)});
+	auto instance = osn::SimpleRecording::constructor.New({Napi::Number::New(info.Env(), static_cast<double>(response[1].value_union.ui64))});
 	return instance;
 }
